Fix 3 test failures: n-bins type, regularization OOM, variance_spatial_fsc
- parser_args.py: --n-bins type=float → type=int (was causing
  TypeError: 'float' object is not iterable in compute_state)
- covariance_estimation.py: regularization batch divisor //4 → //8
  (peak memory ~10 vol-sized arrays per column, not 4; prevents OOM
  when GPU memory is shared with parent process)
- Add spatial_fsc to HIGH_VARIANCE_TOKENS (GPU non-determinism in
  covariance estimation causes ~1% jitter in variance FSC)

+++ recovar/utils/parser_args.py
@@ -32,7 +32,7 @@ def standard_downstream_args(parser: argparse.ArgumentParser, analyze= False):
     )
 
     parser.add_argument(
-        "--n-bins",  type =float, default=50, dest="n_bins",help="number of bins for kernel regression. Default is 50 and works well for most cases. E.g., it was used to generate all figures in the paper"
+        "--n-bins",  type =int, default=50, dest="n_bins",help="number of bins for kernel regression. Default is 50 and works well for most cases. E.g., it was used to generate all figures in the paper"
     )   
 
     parser.add_argument(
